fixing issues with hosting
bluehost is a pain in the ass. 

first off I needed to change some db configuration. this caused issues with emoji's so now they will be handled through the model rather than being stored directly to the db. there was also an issue with key length so I updated the app service provider.